requirements.txt
asgiref==3.8.1
cffi==1.17.1
crispy-bootstrap5==2025.4
cryptography==45.0.3
Django==5.2.1
django-crispy-forms==2.4
django-environ==0.12.0
django-extensions==4.1
pan-os-python==1.12.2
pan-python==0.17.0
pandevice==0.14.0
pycparser==2.22
setuptools==80.9.0
sqlparse==0.5.3
whitenoise==6.9.0
